change type of weight vector

+++ cpp/subprojects/boosting/include/boosting/statistics/statistics_example_wise_common.hpp
@@ -74,7 +74,7 @@ namespace boosting {
                         delete totalCoverableSumVector_;
                     }
 
-                    void addToMissing(uint32 statisticIndex, uint32 weight) override {
+                    void addToMissing(uint32 statisticIndex, float64 weight) override {
                         // Create a vector for storing the totals sums of gradients and Hessians, if necessary...
                         if (totalCoverableSumVector_ == nullptr) {
                             totalCoverableSumVector_ = new StatisticVector(*totalSumVector_);
@@ -90,7 +90,7 @@ namespace boosting {
                             statistics_.statisticMatrixPtr_->hessians_row_cend(statisticIndex), weight);
                     }
 
-                    void addToSubset(uint32 statisticIndex, uint32 weight) override {
+                    void addToSubset(uint32 statisticIndex, float64 weight) override {
                         sumVector_.addToSubset(statistics_.statisticMatrixPtr_->gradients_row_cbegin(statisticIndex),
                                                statistics_.statisticMatrixPtr_->gradients_row_cend(statisticIndex),
                                                statistics_.statisticMatrixPtr_->hessians_row_cbegin(statisticIndex),
@@ -230,7 +230,7 @@ namespace boosting {
 
             }
 
-            void removeFromBin(uint32 binIndex, uint32 statisticIndex, uint32 weight) override {
+            void removeFromBin(uint32 binIndex, uint32 statisticIndex, float64 weight) override {
                 this->statisticMatrixPtr_->subtractFromRow(
                     binIndex, originalStatisticMatrix_.gradients_row_cbegin(statisticIndex),
                     originalStatisticMatrix_.gradients_row_cend(statisticIndex),
@@ -301,7 +301,7 @@ namespace boosting {
                     return statisticMatrixPtr_->getNumRows();
                 }
 
-                void addToBin(uint32 binIndex, uint32 statisticIndex, uint32 weight) override {
+                void addToBin(uint32 binIndex, uint32 statisticIndex, float64 weight) override {
                     statisticMatrixPtr_->addToRow(binIndex,
                                                   statistics_.statisticMatrixPtr_->gradients_row_cbegin(statisticIndex),
                                                   statistics_.statisticMatrixPtr_->gradients_row_cend(statisticIndex),
@@ -376,7 +376,7 @@ namespace boosting {
                 this->resetCoveredStatistics();
             }
 
-            void addSampledStatistic(uint32 statisticIndex, uint32 weight) override {
+            void addSampledStatistic(uint32 statisticIndex, float64 weight) override {
                 // This function is equivalent to the function `updateCoveredStatistic`...
                 this->updateCoveredStatistic(statisticIndex, weight, false);
             }
@@ -385,7 +385,7 @@ namespace boosting {
                 totalSumVectorPtr_->setAllToZero();
             }
 
-            void updateCoveredStatistic(uint32 statisticIndex, uint32 weight, bool remove) override {
+            void updateCoveredStatistic(uint32 statisticIndex, float64 weight, bool remove) override {
                 float64 signedWeight = remove ? -((float64) weight) : weight;
                 totalSumVectorPtr_->add(this->statisticMatrixPtr_->gradients_row_cbegin(statisticIndex),
                                         this->statisticMatrixPtr_->gradients_row_cend(statisticIndex),
