+++ cpp/subprojects/common/include/common/sampling/weight_vector_equal.hpp
@@ -24,8 +24,8 @@ class EqualWeightVector final : public IWeightVector {
 
         bool hasZeroWeights() const override;
 
-        uint32 getWeight(uint32 pos) const override;
+        float64 getWeight(uint32 pos) const override;
 
-        uint32 getSumOfWeights() const override;
+        float64 getSumOfWeights() const override;
 
 };
