+++ cpp/subprojects/common/include/common/statistics/histogram.hpp
@@ -23,6 +23,6 @@ class IHistogram : virtual public IImmutableStatistics {
           * @param statisticIndex   The index of the statistic
           * @param weight           The weight of the statistic
           */
-        virtual void removeFromBin(uint32 binIndex, uint32 statisticIndex, uint32 weight) = 0;
+        virtual void removeFromBin(uint32 binIndex, uint32 statisticIndex, float64 weight) = 0;
 
 };
