+++ cpp/subprojects/common/src/common/thresholds/thresholds_exact.cpp
@@ -122,7 +122,7 @@ static inline void filterCurrentVector(const FeatureVector& vector, FilteredCach
             coverageMaskIterator[index] = numConditions;
             filteredIterator[i].index = index;
             filteredIterator[i].value = iterator[r].value;
-            uint32 weight = weights.getWeight(index);
+            float64 weight = weights.getWeight(index);
             statistics.updateCoveredStatistic(index, weight, false);
             i++;
         }
@@ -156,7 +156,7 @@ static inline void filterCurrentVector(const FeatureVector& vector, FilteredCach
         for (intp r = start; r < end; r++) {
             uint32 index = iterator[r].index;
             coverageMaskIterator[index] = numConditions;
-            uint32 weight = weights.getWeight(index);
+            float64 weight = weights.getWeight(index);
             statistics.updateCoveredStatistic(index, weight, true);
         }
 
@@ -187,7 +187,7 @@ static inline void filterCurrentVector(const FeatureVector& vector, FilteredCach
         for (auto it = vector.missing_indices_cbegin(); it != vector.missing_indices_cend(); it++) {
             uint32 index = *it;
             coverageMaskIterator[index] = numConditions;
-            uint32 weight = weights.getWeight(index);
+            float64 weight = weights.getWeight(index);
             statistics.updateCoveredStatistic(index, weight, true);
         }
     }
@@ -321,7 +321,7 @@ class ExactThresholds final : public AbstractThresholds {
 
             const IWeightVector& weights_;
 
-            uint32 sumOfWeights_;
+            float64 sumOfWeights_;
 
             CoverageMask coverageMask_;
